langgraph-agent/requirements.txt
# ============================================================================
# LangGraph Interview Agent - Dependencies (GOOGLE GEMINI VERSION)
# ============================================================================

# Core LangGraph & LangChain (sin versiones fijas, pip resuelve)
langgraph>=0.2.0,<0.3.0
langchain>=0.3.0,<0.4.0
langchain-google-genai>=2.0.0,<3.0.0
langchain-core>=0.3.0,<0.4.0

# Google AI
google-generativeai>=0.8.0,<0.9.0

# FastAPI & Server
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.9.0
pydantic-settings>=2.5.0

# Database - Supabase
supabase>=2.8.0
postgrest>=0.17.0

# HTTP Clients
httpx>=0.27.0
requests>=2.32.0

# Environment & Config
python-dotenv>=1.0.0

# Date/Time utilities
python-dateutil>=2.9.0

# Structured Logging
structlog>=24.4.0

# Development & Testing (opcional)
pytest>=8.3.0
pytest-asyncio>=0.24.0
black>=24.8.0
flake8>=7.1.0
mypy>=1.11.0
